requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.1
django-allauth==0.44.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-phonenumber-field==5.1.0
django-simpleui==2021.4.3
django-split-settings==1.0.1
idna==2.10
oauthlib==3.1.0
phonenumbers==8.12.22
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
